+++ cli/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Auto-create collections when serving ([#304](https://github.com/stac-utils/stac-rs/pull/304))
 - Auto-add items when serving ([#312](https://github.com/stac-utils/stac-rs/pull/312))
 - Searching geoparquet files with DuckDB ([#331](https://github.com/stac-utils/stac-rs/pull/331))
+- Python package ([#335](https://github.com/stac-utils/stac-rs/pull/335))
 
 ## [0.2.0] - 2024-08-12